Increase global scaling on DP-1 to 150%
We got complaints that the font on our 4k monitors is hard to read
because it's too small. Our FullHD monitors are all connected via VGA
and the beamer is connected on DP-2 on the PC that is connected to the
beamer, so just configuring a different scaling for DP-1 should work.